requirements.txt
fastapi
pydantic
streamlit
requests
openai
uvicorn 
python-dotenv